Rename reference_count files to reference_counter files
Previously, files associated with the ReferenceCounter
class were named reference_count.*. This is
inconsistent with the remainder of the repo as the file
typical reflect the class it contains.
We rename to reference counter for consistency.
